docs: refactor starter story for tests
