+++ examples/react-webpack/.config/buildtime.js
@@ -17,7 +17,7 @@ module.exports = {
     '../../stories/src/tutorial/testing/*.mdx',
     '../../stories/src/tutorial/configuration/*.mdx',
     '../../stories/src/tutorial/reference/*.mdx',
-    '../../stories/src/stories/*.stories.@(js|jsx|tsx|mdx)',
+    '../../stories/src/stories/**/*.stories.@(js|jsx|tsx|mdx)',
     '../../stories/src/stories_native/*.stories.@(js|jsx|tsx|mdx)',
     '../../stories/src/mdx-stories/*.mdx',
     '../../stories/src/catalogs/*.mdx',
