+++ examples/stories/.config/buildtime.js
@@ -1,0 +1,5 @@
+module.exports = {
+  stories: [
+    '../src/**/*.stories.tsx',
+  ],
+};
