+++ examples/stories/src/stories/VariantButton/media/image_screenshot.jpg
