+++ ui/blocks/src/ComponentJSX/ComponentJSX.tsx
@@ -20,6 +20,7 @@ const NAME = 'component_jsx';
 export const ComponentJSX: FC<ComponentJSXProps> = fullProps => {
   const props = useCustomProps<ComponentJSXProps>(NAME, fullProps);
   const component = useComponent({ of: props.of });
+  console.log(component);
   if (!component?.jsx?.length) {
     return null;
   }
